+++ keyboards/hwpm87/config.h
@@ -10,7 +10,7 @@
 #define PRODUCT_ID      0x0000
 #define DEVICE_VER      0x0001
 #define MANUFACTURER    KD-MM2
-#define PRODUCT         promicro87
+#define PRODUCT         HWPM87
 
 /* key matrix size */
 #define MATRIX_ROWS 6
