+++ keyboards/hwpm87/hwpm87.c
@@ -1,4 +1,4 @@
 // Copyright 2022 CAO THAI DUONG (@KD-MM2)
 // SPDX-License-Identifier: GPL-2.0-or-later
 
-#include "promicro87.h"
+#include "hwpm87.h"
